backend/auth-yandex/requirements.txt
pyjwt==2.8.0
requests==2.31.0
psycopg2-binary==2.9.9
